requirements.txt
pytesseract~=0.3.10
Pillow~=9.3.0
numpy~=1.23.4